friends are targeted correctly
